move NPCResults into tests, and remove its use in production code. We have SpendBundleConditions which captures the non-error state of NPCResult
