+++ chia/_tests/clvm/benchmark_costs.py
@@ -3,8 +3,7 @@ from __future__ import annotations
 from chia_rs import SpendBundle
 from chia_rs.sized_ints import uint64
 
-from chia._tests.util.get_name_puzzle_conditions import get_name_puzzle_conditions
-from chia.consensus.cost_calculator import NPCResult
+from chia._tests.util.get_name_puzzle_conditions import NPCResult, get_name_puzzle_conditions
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
 from chia.full_node.bundle_tools import simple_solution_generator
 from chia.types.blockchain_format.program import INFINITE_COST
